% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{crr_optionvalue}
\alias{crr_optionvalue}
\title{Compute option values}
\usage{
crr_optionvalue(
  S0,
  u,
  d,
  r,
  expiry,
  implicit_value,
  american = FALSE,
  simple = TRUE,
  geometric = TRUE
)
}
\arguments{
\item{S0}{The initial stock price}

\item{u}{The up factor}

\item{d}{The down factor}

\item{r}{The risk-free interest rate}

\item{expiry}{The time to expiry in discrete steps}

\item{implicit_value}{A function which takes a price matrix and returns a
vector of option values. The last column in the matrix is taken to be the
'current' price. This is usually the result of calling \code{implicit_value()}.}

\item{american}{A logical, whether the contract allows for early exercise}

\item{simple}{A logical whether to use simple interest or continuously
compounding interest}

\item{geometric}{A logical, whether the stock price movements are geometric
or arithmetic. Note that arithmetic movements can currently only be used if
\code{r=0}.}
}
\value{
A list containing: (i) a path matrix, in which each row corresponds
to a path, each column corresponds to a time-step, zeroes correspond to
down moves and ones to up-moves; (ii) a corresponding price matrix,
containing the stock price at every node in the lattice; and (iii) a
corresponding value matrix, containing the option value at every node in
the lattice
}
\description{
Compute option values
}
\examples{

# Vanilla European call:
crr_optionvalue(
  S0 = 62,
  u = 1.05943,
  d = 1/1.05943,
  r = 0.1/12, # 10\% p.a compounded monthly
  expiry = 5,
  implicit_value = implicit_value(Sk=60, contract = "call"),
  american = FALSE
)

# Asian European call:
crr_optionvalue(
  S0 = 4,
  u = 2,
  d = 1/2,
  r = 1/4,
  expiry = 3,
  implicit_value = implicit_value(Sk=4, contract = "call", exotic = "asian"),
  american = FALSE
)

}
