% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{crr_plot}
\alias{crr_plot}
\title{Plot stock prices and option values}
\usage{
crr_plot(res)
}
\arguments{
\item{res}{The result of either \code{crr_stockprice} or \code{crr_optionvalue}}
}
\value{
A ggplot object
}
\description{
Plot stock prices (blue) and option values (red) for each 'position' in the
lattice, where a 'position' is determined entirely by the number of up- and
down-jumps taken to get there. Note that this means that, e.g., the up-down
and down-up nodes will be displayed on top of one another. Consequently,
multiple value labels will attach to each position when exotic options
are being priced.
}
\examples{
res <- crr_optionvalue(
  S0 = 62,
  u = 1.05943,
  d = 1/1.05943,
  r = 0.1/12, # 10\% p.a compounded monthly
  expiry = 5,
  implicit_value = implicit_value(Sk=60, contract = "call"),
  american = FALSE
)

crr_plot(res)
}
