% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{crr_stockprice}
\alias{crr_stockprice}
\title{Compute stock prices}
\usage{
crr_stockprice(S0, u, d, expiry, geometric = TRUE)
}
\arguments{
\item{S0}{The initial stock price}

\item{u}{The up factor}

\item{d}{The down factor}

\item{expiry}{The time to expiry in discrete steps}

\item{geometric}{A logical, whether the stock price movements are geometric
or arithmetic. Note that arithmetic movements can currently only be used if
\code{r=0}.}
}
\value{
A list containing: (i) a path matrix, in which each row corresponds
to a path, each column corresponds to a time-step, zeroes correspond to
down moves and ones to up-moves; and (ii) a corresponding price matrix,
containing the stock price at every node in the lattice.
}
\description{
Compute stock prices
}
\examples{
res <- crr_stockprice(
  S0 = 62,
  u = 1.05943,
  d = 1/1.05943,
  expiry = 5
)
crr_plot(res)

}
