% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoff.R
\name{implicit_value}
\alias{implicit_value}
\title{Payoff function factory}
\usage{
implicit_value(
  Sk,
  Sb = NULL,
  contract = c("call", "put", "straddle"),
  exotic = c("no", "lookback_max", "lookback_min", "asian"),
  barrier = c("none", "ui", "uo", "di", "do")
)
}
\arguments{
\item{Sk}{The strike price}

\item{Sb}{The barrier price (ignored if \code{barrier='none'})}

\item{contract}{The contract type}

\item{exotic}{The variety of 'exoticness'}

\item{barrier}{The type of barrier, up-and-in ('ui'), down-and-out ('do'), etc.}
}
\value{
A function which takes a price matrix and returns a vector of option
values.
}
\description{
Payoff function factory
}
