% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{S0}{The initial stock price}

\item{Sk}{The strike price}

\item{Sb}{The barrier price (ignored if \code{barrier='none'})}

\item{contract}{The contract type}

\item{exotic}{The variety of 'exoticness'}

\item{barrier}{The type of barrier, up-and-in ('ui'), down-and-out ('do'), etc.}

\item{u}{The up factor}

\item{d}{The down factor}

\item{r}{The risk-free interest rate}

\item{expiry}{The time to expiry in discrete steps}

\item{simple}{A logical whether to use simple interest or continuously
compounding interest}

\item{implicit_value}{A function which takes a price matrix and returns a
vector of option values. The last column in the matrix is taken to be the
'current' price. This is usually the result of calling \code{implicit_value()}.}

\item{american}{A logical, whether the contract allows for early exercise}

\item{geometric}{A logical, whether the stock price movements are geometric
or arithmetic. Note that arithmetic movements can currently only be used if
\code{r=0}.}
}
\description{
Parameters
}
